Time: 3 ms (21.74%) | Memory: 7.5 MB (54.15%) - LeetSync

+++ 440-k-th-smallest-in-lexicographical-order/k-th-smallest-in-lexicographical-order.cpp
@@ -4,7 +4,7 @@ public:
     int findKthNumber(int n, int k) {
         int curr=1;
         k-=1;
-        long step;
+        int step;
 
         while(k>0)
         {
